Check if joining the wifi was successful...
...and exit accordingly

+++ application/senic/nuimo_hub/commands.py
@@ -3,7 +3,7 @@ import json
 import os
 
 from os.path import abspath
-from subprocess import run
+from subprocess import run, PIPE
 
 import wifi
 from pyramid.paster import get_app
@@ -86,6 +86,15 @@ def join_wifi(ssid, password, device=DEFAULT_IFACE):
     with open(WPA_SUPPLICANT_FS, 'w') as wpaconf:
         wpaconf.write(WPA_SUPPLICANT_CONF.format(**locals()))
     run(['ifup', device])
-    # clean up after ourselves
-    if os.path.exists(ENTER_SETUP_FLAG):
-        os.remove(ENTER_SETUP_FLAG)
+
+    # check, if we were successful:
+    status = run(['wpa_cli', 'status'], stdout=PIPE)
+    success = 'wpa_state=COMPLETED' in status.stdout.decode()
+
+    if success:
+        # clean up after ourselves
+        if os.path.exists(ENTER_SETUP_FLAG):
+            os.remove(ENTER_SETUP_FLAG)
+        exit(0)
+    else:
+        exit(1)
